requirements.txt
numpy~=1.26.4
pandas~=2.3.2
torch~=2.8.0+cpu
matplotlib~=3.10.5
scikit-learn~=1.7.1
tensorflow~=2.20.0
tqdm~=4.67.1
python-dotenv~=1.1.1
gensim~=4.3.3
lxml~=6.0.1